Add standalone class for Place App Domain Event
Added a standalone class for managing placement application domain events
