+++ src/main/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/service/PlacementApplicationService.kt
@@ -31,6 +31,7 @@ import uk.gov.justice.digital.hmpps.approvedpremisesapi.model.validated
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.problem.InternalServerErrorProblem
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.results.AuthorisableActionResult
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.results.ValidatableActionResult
+import uk.gov.justice.digital.hmpps.approvedpremisesapi.service.cas1.Cas1PlacementApplicationDomainEventService
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.service.cas1.Cas1PlacementApplicationEmailService
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.transformer.DomainEventTransformer
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.util.UrlTemplate
@@ -56,8 +57,7 @@ class PlacementApplicationService(
   private val notifyConfig: NotifyConfig,
   private val userAccessService: UserAccessService,
   private val cas1PlacementApplicationEmailService: Cas1PlacementApplicationEmailService,
-  private val domainEventService: DomainEventService,
-  private val domainEventTransformer: DomainEventTransformer,
+  private val cas1PlacementApplicationDomainEventService: Cas1PlacementApplicationDomainEventService,
   @Value("\${notify.send-placement-request-notifications}")
   private val sendPlacementRequestNotifications: Boolean,
   @Value("\${url-templates.frontend.application}") private val applicationUrlTemplate: UrlTemplate,
@@ -228,52 +228,13 @@ class PlacementApplicationService(
 
     val savedApplication = placementApplicationRepository.save(placementApplication)
 
-    createWithdrawalDomainEvent(placementApplication, user)
-
+    cas1PlacementApplicationDomainEventService.placementApplicationWithdrawn(placementApplication, withdrawalContext)
     cas1PlacementApplicationEmailService.placementApplicationWithdrawn(placementApplication, wasBeingAssessedBy)
 
     withdrawPlacementRequests(placementApplication, withdrawalContext)
 
     return AuthorisableActionResult.Success(
       ValidatableActionResult.Success(savedApplication),
-    )
-  }
-
-  private fun createWithdrawalDomainEvent(
-    placementApplication: PlacementApplicationEntity,
-    user: UserEntity,
-  ) {
-    val domainEventId = UUID.randomUUID()
-    val eventOccurredAt = Instant.now()
-    val application = placementApplication.application
-
-    val placementApplicationWithdrawn = PlacementApplicationWithdrawn(
-      applicationId = application.id,
-      applicationUrl = applicationUrlTemplate.resolve("id", application.id.toString()),
-      placementApplicationId = placementApplication.id,
-      personReference = PersonReference(
-        crn = application.crn,
-        noms = application.nomsNumber ?: "Unknown NOMS Number",
-      ),
-      deliusEventNumber = application.eventNumber,
-      withdrawnAt = eventOccurredAt,
-      withdrawnBy = domainEventTransformer.toWithdrawnBy(user),
-      withdrawalReason = placementApplication.withdrawalReason!!.name,
-    )
-
-    domainEventService.savePlacementApplicationWithdrawnEvent(
-      DomainEvent(
-        id = domainEventId,
-        applicationId = application.id,
-        crn = application.crn,
-        occurredAt = eventOccurredAt,
-        data = PlacementApplicationWithdrawnEnvelope(
-          id = domainEventId,
-          timestamp = eventOccurredAt,
-          eventType = "approved-premises.placement-application.withdrawn",
-          eventDetails = placementApplicationWithdrawn,
-        ),
-      ),
     )
   }
