+++ src/test/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/unit/service/PlacementApplicationServiceTest.kt
@@ -55,6 +55,7 @@ import uk.gov.justice.digital.hmpps.approvedpremisesapi.service.UserAccessServic
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.service.UserService
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.service.WithdrawableEntityType
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.service.WithdrawalContext
+import uk.gov.justice.digital.hmpps.approvedpremisesapi.service.cas1.Cas1PlacementApplicationDomainEventService
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.service.cas1.Cas1PlacementApplicationEmailService
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.transformer.DomainEventTransformer
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.util.UrlTemplate
@@ -73,9 +74,8 @@ class PlacementApplicationServiceTest {
   private val userAllocator = mockk<UserAllocator>()
   private val notifyConfig = mockk<NotifyConfig>()
   private val userAccessService = mockk<UserAccessService>()
-  private val domainEventService = mockk<DomainEventService>()
-  private val domainEventTransformer = mockk<DomainEventTransformer>()
   private val cas1PlacementApplicationEmailService = mockk<Cas1PlacementApplicationEmailService>()
+  private val cas1PlacementApplicationDomainEventService = mockk<Cas1PlacementApplicationDomainEventService>()
 
   private val placementApplicationService = PlacementApplicationService(
     placementApplicationRepository,
@@ -88,8 +88,7 @@ class PlacementApplicationServiceTest {
     notifyConfig,
     userAccessService,
     cas1PlacementApplicationEmailService,
-    domainEventService,
-    domainEventTransformer,
+    cas1PlacementApplicationDomainEventService,
     sendPlacementRequestNotifications = true,
     UrlTemplate("http://frontend/applications/#id"),
   )
@@ -566,7 +565,7 @@ class PlacementApplicationServiceTest {
     }
 
     @Test
-    fun `it withdraws a placement application and triggers emails`() {
+    fun `it withdraws a placement application and triggers emails and domain events`() {
       val allocatedTo = UserEntityFactory().withDefaultProbationRegion().produce()
 
       val placementApplication = PlacementApplicationEntityFactory()
@@ -577,23 +576,23 @@ class PlacementApplicationServiceTest {
         .produce()
 
       val templateId = UUID.randomUUID().toString()
+      val withdrawalContext = WithdrawalContext(
+        user,
+        WithdrawableEntityType.PlacementApplication
+      )
 
       every { placementApplicationRepository.findByIdOrNull(placementApplication.id) } returns placementApplication
       every { userAccessService.userMayWithdrawPlacementApplication(any(), any()) } returns true
       every { notifyConfig.templates.placementRequestWithdrawn } answers { templateId }
       every { emailNotificationService.sendEmail(any(), any(), any()) } returns Unit
       every { placementApplicationRepository.save(any()) } answers { it.invocation.args[0] as PlacementApplicationEntity }
-      every { domainEventTransformer.toWithdrawnBy(user) } returns WithdrawnByFactory().produce()
-      every { domainEventService.savePlacementApplicationWithdrawnEvent(any()) } returns Unit
+      every { cas1PlacementApplicationDomainEventService.placementApplicationWithdrawn(placementApplication,withdrawalContext) } returns Unit
       every { cas1PlacementApplicationEmailService.placementApplicationWithdrawn(any(), any()) } returns Unit
 
       val result = placementApplicationService.withdrawPlacementApplication(
         placementApplication.id,
         PlacementApplicationWithdrawalReason.ALTERNATIVE_PROVISION_IDENTIFIED,
-        withdrawalContext = WithdrawalContext(
-          user,
-          WithdrawableEntityType.PlacementApplication,
-        ),
+        withdrawalContext = withdrawalContext,
       )
 
       assertThat(result is AuthorisableActionResult.Success).isTrue
@@ -607,61 +606,6 @@ class PlacementApplicationServiceTest {
       assertThat(entity.decision).isEqualTo(PlacementApplicationDecision.WITHDRAW)
 
       verify { cas1PlacementApplicationEmailService.placementApplicationWithdrawn(placementApplication, allocatedTo) }
-    }
-
-    @Test
-    fun `it creates a domain event`() {
-      val placementApplication = PlacementApplicationEntityFactory()
-        .withApplication(application)
-        .withAllocatedToUser(UserEntityFactory().withDefaultProbationRegion().produce())
-        .withDecision(null)
-        .withCreatedByUser(user)
-        .produce()
-
-      val templateId = UUID.randomUUID().toString()
-
-      every { placementApplicationRepository.findByIdOrNull(placementApplication.id) } returns placementApplication
-      every { userAccessService.userMayWithdrawPlacementApplication(any(), any()) } returns true
-      every { notifyConfig.templates.placementRequestWithdrawn } answers { templateId }
-      every { emailNotificationService.sendEmail(any(), any(), any()) } returns Unit
-      every { placementApplicationRepository.save(any()) } answers { it.invocation.args[0] as PlacementApplicationEntity }
-      every { cas1PlacementApplicationEmailService.placementApplicationWithdrawn(any(), any()) } returns Unit
-
-      val withdrawnBy = WithdrawnByFactory().produce()
-      every { domainEventTransformer.toWithdrawnBy(user) } returns withdrawnBy
-      every { domainEventService.savePlacementApplicationWithdrawnEvent(any()) } returns Unit
-
-      val result = placementApplicationService.withdrawPlacementApplication(
-        placementApplication.id,
-        PlacementApplicationWithdrawalReason.ALTERNATIVE_PROVISION_IDENTIFIED,
-        withdrawalContext = WithdrawalContext(
-          user,
-          WithdrawableEntityType.PlacementApplication,
-        ),
-      )
-
-      assertThat(result is AuthorisableActionResult.Success).isTrue
-
-      verify(exactly = 1) {
-        domainEventService.savePlacementApplicationWithdrawnEvent(
-          match {
-            val data = it.data.eventDetails
-
-            it.applicationId == application.id &&
-              it.crn == application.crn &&
-              data.applicationId == application.id &&
-              data.applicationUrl == "http://frontend/applications/${application.id}" &&
-              data.placementApplicationId == placementApplication.id &&
-              data.personReference == PersonReference(
-              crn = application.crn,
-              noms = application.nomsNumber!!,
-            ) &&
-              data.deliusEventNumber == application.eventNumber &&
-              data.withdrawnBy == withdrawnBy
-            data.withdrawalReason == "ALTERNATIVE_PROVISION_IDENTIFIED"
-          },
-        )
-      }
     }
 
     @Test
@@ -679,8 +623,7 @@ class PlacementApplicationServiceTest {
       every { notifyConfig.templates.placementRequestWithdrawn } answers { templateId }
       every { emailNotificationService.sendEmail(any(), any(), any()) } returns Unit
       every { placementApplicationRepository.save(any()) } answers { it.invocation.args[0] as PlacementApplicationEntity }
-      every { domainEventTransformer.toWithdrawnBy(user) } returns WithdrawnByFactory().produce()
-      every { domainEventService.savePlacementApplicationWithdrawnEvent(any()) } returns Unit
+      every { cas1PlacementApplicationDomainEventService.placementApplicationWithdrawn(any(), any()) } returns Unit
       every { cas1PlacementApplicationEmailService.placementApplicationWithdrawn(any(), any()) } returns Unit
 
       val result = placementApplicationService.withdrawPlacementApplication(
@@ -756,8 +699,7 @@ class PlacementApplicationServiceTest {
       every { notifyConfig.templates.placementRequestWithdrawn } answers { templateId }
       every { emailNotificationService.sendEmail(any(), any(), any()) } returns Unit
       every { placementApplicationRepository.save(any()) } answers { it.invocation.args[0] as PlacementApplicationEntity }
-      every { domainEventTransformer.toWithdrawnBy(user) } returns WithdrawnByFactory().produce()
-      every { domainEventService.savePlacementApplicationWithdrawnEvent(any()) } returns Unit
+      every { cas1PlacementApplicationDomainEventService.placementApplicationWithdrawn(any(), any()) } returns Unit
       every { cas1PlacementApplicationEmailService.placementApplicationWithdrawn(any(), any()) } returns Unit
       every {
         placementRequestService.withdrawPlacementRequest(any(), any(), any())
@@ -819,8 +761,7 @@ class PlacementApplicationServiceTest {
       every { notifyConfig.templates.placementRequestWithdrawn } answers { templateId }
       every { emailNotificationService.sendEmail(any(), any(), any()) } returns Unit
       every { placementApplicationRepository.save(any()) } answers { it.invocation.args[0] as PlacementApplicationEntity }
-      every { domainEventTransformer.toWithdrawnBy(user) } returns WithdrawnByFactory().produce()
-      every { domainEventService.savePlacementApplicationWithdrawnEvent(any()) } returns Unit
+      every { cas1PlacementApplicationDomainEventService.placementApplicationWithdrawn(any(), any()) } returns Unit
       every { cas1PlacementApplicationEmailService.placementApplicationWithdrawn(any(), any()) } returns Unit
       every {
         placementRequestService.withdrawPlacementRequest(any(), any(), any())
